refactor(settings): separate app panel to components
