+++ app/react/portainer/settings/SettingsView/ApplicationSettingsPanel/index.ts
@@ -1,0 +1,1 @@
+export { ApplicationSettingsPanel } from './ApplicationSettingsPanel';
