+++ app/react/portainer/settings/SettingsView/ApplicationSettingsPanel/types.ts
@@ -1,0 +1,10 @@
+export interface Values {
+  snapshotInterval: string;
+  edgeAgentCheckinInterval: number;
+  enableTelemetry: boolean;
+  loginBanner: string;
+  loginBannerEnabled: boolean;
+  logo: string;
+  logoEnabled: boolean;
+  templatesUrl: string;
+}
